Finalizing removal structure. Now it creates a .<packagename> file in .ccp4mpkgs/ which tells what to remove
