+++ include/PackageManagement/PackageInstaller.hpp
@@ -9,6 +9,6 @@ bool InstallDirectory( const Package & pkg );
 
 void GetCopyCommands( const Package & pkg, std::string & include, std::vector< std::string > & libs, std::string & framework );
 
-bool RemoveAllCopiedFiles( std::vector< std::string > & files );
+void RevertInstallation( std::vector< std::string > & copiedfiles );
 
 #endif // PACKAGEINSTALLER_HPP
